move Rules impl, fix tests
