+++ java/core/src/main/java/co/worklytics/psoxy/rules/Rules1.java
@@ -1,6 +1,5 @@
-package co.worklytics.psoxy;
+package co.worklytics.psoxy.rules;
 
-import co.worklytics.psoxy.rules.RuleSet;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 import lombok.*;
