+++ java/core/src/main/java/co/worklytics/psoxy/storage/StorageHandler.java
@@ -1,11 +1,11 @@
 package co.worklytics.psoxy.storage;
 
-import co.worklytics.psoxy.Rules1;
 import co.worklytics.psoxy.Sanitizer;
 import co.worklytics.psoxy.SanitizerFactory;
 import co.worklytics.psoxy.gateway.ConfigService;
 import co.worklytics.psoxy.gateway.StorageEventRequest;
 import co.worklytics.psoxy.gateway.StorageEventResponse;
+import co.worklytics.psoxy.rules.RuleSet;
 import lombok.NoArgsConstructor;
 import lombok.SneakyThrows;
 
@@ -17,7 +17,7 @@ public class StorageHandler {
     @Inject
     ConfigService config;
     @Inject
-    Rules1 rules1;
+    RuleSet rules;
     @Inject
     FileHandlerFactory fileHandlerStrategy;
     @Inject
@@ -30,7 +30,7 @@ public class StorageHandler {
         if (this.sanitizer == null) {
             synchronized ($writeLock) {
                 if (this.sanitizer == null) {
-                    this.sanitizer = sanitizerFactory.create(sanitizerFactory.buildOptions(config, rules1));
+                    this.sanitizer = sanitizerFactory.create(sanitizerFactory.buildOptions(config, rules));
                 }
             }
         }
