+++ java/core/src/test/java/co/worklytics/psoxy/rules/JavaRulesTestBaseCase.java
@@ -1,7 +1,5 @@
 package co.worklytics.psoxy.rules;
 
-import co.worklytics.psoxy.Rules1;
-import co.worklytics.psoxy.Rules2;
 import co.worklytics.psoxy.rules.google.PrebuiltSanitizerRules;
 import lombok.SneakyThrows;
 import org.junit.jupiter.api.Test;
