+++ java/core/src/test/java/co/worklytics/psoxy/rules/google/Calendar2Tests.java
@@ -1,7 +1,7 @@
 
 package co.worklytics.psoxy.rules.google;
 
-import co.worklytics.psoxy.Rules2;
+import co.worklytics.psoxy.rules.Rules2;
 import co.worklytics.psoxy.rules.JavaRules2TestBaseCase;
 import lombok.Getter;
 import lombok.SneakyThrows;
