+++ java/core/src/test/java/co/worklytics/psoxy/rules/google/DirectoryTests.java
@@ -1,6 +1,6 @@
 package co.worklytics.psoxy.rules.google;
 
-import co.worklytics.psoxy.Rules1;
+import co.worklytics.psoxy.rules.Rules1;
 import co.worklytics.psoxy.Sanitizer;
 import co.worklytics.psoxy.rules.JavaRulesTestBaseCase;
 import lombok.Getter;
