+++ java/core/src/test/java/co/worklytics/psoxy/rules/msft/CalendarTests.java
@@ -1,7 +1,6 @@
 package co.worklytics.psoxy.rules.msft;
 
-import co.worklytics.psoxy.Rules1;
-import co.worklytics.psoxy.Rules2;
+import co.worklytics.psoxy.rules.Rules2;
 import lombok.Getter;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
