+++ java/core/src/test/java/co/worklytics/psoxy/rules/slack/SlackDiscoveryTests.java
@@ -1,6 +1,6 @@
 package co.worklytics.psoxy.rules.slack;
 
-import co.worklytics.psoxy.Rules1;
+import co.worklytics.psoxy.rules.Rules1;
 import co.worklytics.psoxy.rules.JavaRulesTestBaseCase;
 import lombok.Getter;
 import lombok.SneakyThrows;
