+++ java/core/src/test/java/co/worklytics/psoxy/storage/impl/CSVFileHandlerTest.java
@@ -1,7 +1,7 @@
 package co.worklytics.psoxy.storage.impl;
 
 import co.worklytics.psoxy.PsoxyModule;
-import co.worklytics.psoxy.Rules1;
+import co.worklytics.psoxy.rules.Rules1;
 import co.worklytics.psoxy.Sanitizer;
 import co.worklytics.psoxy.SanitizerFactory;
 import co.worklytics.psoxy.gateway.ConfigService;
